+++ CHANGELOG.md
@@ -37,3 +37,4 @@
 ## 0.3.0
 - Fix problem with Bible Fact Pak™ import where existing items will be re-imported a second time
 - Move Bible Fact Pak™ import user interface into a user control
+- Refactor IQuestionImporter interface so default constructors can now be used